// projects/app/src/app/interfaces/user.ts
export interface User {
  stripeId?: string;
  stripeLink?: string;
}

export interface OandaCredentials {
  apiKey?: string;
  accountId?: string;
}

export interface Settings {
  bot: {
    status: BotStatus;
    notifications: BotNotificationSettings;
  };
  oanda: {
    credentials: OandaCredentials;
    notifications: OandaNotificationSettings;
  };
}

interface BotNotificationSettings {
  onStart: boolean;
  onStop: boolean;
  onPing: boolean;
}

interface OandaNotificationSettings {
  orders: {
    ordersCancelled: boolean;
    ordersCreated: boolean;
    ordersFilled: boolean;
    ordersTriggered: boolean;
  };
  trades: {
    tradesClosed: boolean;
    tradesOpened: boolean;
    tradesReduced: boolean;
  };
}